[SPARK-50974][ML][PYTHON][CONNECT] Add support foldCol for CrossValidator on connect
### What changes were proposed in this pull request?

This PR adds support foldCol for CrossValidator on connect

### Why are the changes needed?
feature parity

### Does this PR introduce _any_ user-facing change?
Yes

### How was this patch tested?
The CI passes

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #49743 from wbo4958/cv-fold.

Authored-by: Bobby Wang <wbo4958@gmail.com>
